move validations and use append instead of +
